Split refunds tests (#18830)
* Split order grant refund create - reasons

* Extract transaction_request_action.py and test_order_grant_refund_update.py

* Fix test file naming: use underscore instead of hyphen

Rename test_transaction_request_action-reason.py to
test_transaction_request_action_reason.py. Python cannot import
modules with hyphens in their names, causing pytest to fail to
discover fixtures defined in the file.


* extract fixture

---------
